requirements.txt
audio_recorder_streamlit==0.0.8
openai==0.27.8
python-dotenv==1.0.1
streamlit==1.31.1
gtts==2.3.1
